fix: added the isMounted pattern to avoid memory leaks
I found that sometimes when searching I would get an error in the console saying that I had a memory leak and I should clean up subscriptions in react components in the componentWillUnmount lifecycle method.

This was the solution I found for that. It appears to just make sure that the component is still actually mounted before attempting to do its async work.