fix: solved the memoyr leak problem
this seemed to be an issue with state and prop management coming down to the MoveBook component. I was storing where the book was stored in the state of the MoveBook component, but that was rather foolish as it would be much better to pass that in as a prop so that when state changes higher up, it will change the MoveBook component as well and not conflict with state stored inside of it, which is what it was doing.

I also converted the MoveBook component into a stateless functional component as I no longer needed to store state in it.